requirements.txt
Django~=4.2.11
djangorestframework~=3.15.1
PyJWT~=1.7.1
psycopg2>=2.8.6,<2.9
